Start a server socket and wait for client request
Start a TCP server socket on ipv4 address and default protocol. Set
socket option SO_REUSEADDR to ON. Bind the socket to port 8080 and
listen for client requests. When client makes a request, accept it using
accept function. At last immediately close the connection without doing
anything.

This will let one client connect to the server and exit without any
response. We can check if the client can successfully connect to server
or not. The client side will report an error because we are not reading
the request.