Read client request in buffer and reply with 200
Create a string buffer of size 2048 and initialize it with 0. Call read
function to read the client request onto the buffer. Send a '200 OK'
reply with no headers and empty body.

The client will now receive a '200 OK' reply regardless of the content
of the request. This will resolve the error appearing on client side
caused by the previous commit.